keyboard sprinkles.

+++ opengl/src/main/java/de/fhwedel/opengl/Column.java
@@ -2,11 +2,11 @@ package de.fhwedel.opengl;
 
 public class Column {
 
-    public double height;
-    public double velocity;
+    public float height;
+    public float velocity;
     public float[] normal;
 
-    public Column(double height, double velocity, float[] normal) {
+    public Column(float height, float velocity, float[] normal) {
         this.height = height;
         this.velocity = velocity;
         this.normal = normal;
